requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
packaging==24.1
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.6
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
